+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## tachyonfx-renderer 0.2.0 - TBD
+- make effects loop automatically with configurable sleep
+
 ## tachyonfx-renderer 0.1.0 - 2025-10-14
 - Initial release